requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.3
fonttools==4.50.0
fsspec==2024.3.1
gradio_client==0.14.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.22.1
idna==3.6
importlib_resources==6.4.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
openai==1.14.3
packaging==24.0
pandas==2.2.1
pillow==10.2.0
PyAudio==0.2.14
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
sympy==1.12
tokenizers==0.15.2
torch==2.2.1
torchaudio==2.2.1
tqdm==4.66.2
transformers==4.39.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Wave==0.0.2
websockets==11.0.3
yml==0.0.1
zipp==3.18.1
